+++ CHANGELOG.md
@@ -1,4 +1,28 @@
-## 0.8.0 (YYYY-MM-DD)
+## 0.9.0 (YYYY-MM-DD)
+
+### Breaking Changes
+* None.
+
+### Enhancements
+* Added API for inspecting the schema of the realm with `BaseRealm.schema()` ([#238](https://github.com/realm/realm-kotlin/issues/238)).
+
+### Fixed
+* None.
+
+### Compatibility
+* This release is compatible with:
+  * Kotlin 1.6.10.
+  * Coroutines 1.5.2-native-mt.
+  * AtomicFu 0.17.0.
+* Minimum Gradle version: 6.1.1.
+* Minimum Android Gradle Plugin version: 4.0.0.
+* Minimum Android SDK: 16.
+
+### Internal
+* None.
+
+
+## 0.8.0 (2021-12-17)
 
 ### Breaking Changes
 * Reworked configuration hierarchy:
@@ -10,6 +34,7 @@
 * [Sync] Added support for `User.logOut()` ([#245](https://github.com/realm/realm-kotlin/issues/245)).
 * Added supported for dates through a new property type: `RealmInstant`.
 * Allow to pass schema as a variable containing the involved `KClass`es and build configurations non-fluently ([#389](https://github.com/realm/realm-kotlin/issues/389)).
+* Added M1 support for `library-base` variant ([#483](https://github.com/realm/realm-kotlin/issues/483)).
 * Added API for inspecting the schema of the realm with `BaseRealm.schema()` ([#238](https://github.com/realm/realm-kotlin/issues/238)).
 
 ### Fixed
@@ -19,9 +44,9 @@
 
 ### Compatibility
 * This release is compatible with:
-  * Kotlin 1.5.31.
+  * Kotlin 1.6.10.
   * Coroutines 1.5.2-native-mt.
-  * AtomicFu 0.16.3.
+  * AtomicFu 0.17.0.
 * Minimum Gradle version: 6.1.1.  
 * Minimum Android Gradle Plugin version: 4.0.0.
 * Minimum Android SDK: 16.
@@ -30,15 +55,17 @@
 * Updated to Ktor 1.6.5.
 * Updated to AndroidX Startup 1.1.0.
 * Updated to Gradle 7.2.
-* Updated to Android Gradle Plugin 7.1.0-beta03.
+* Updated to Android Gradle Plugin 7.1.0-beta05.
 * Updated to NDK 23.1.7779620.
 * Updated to Android targetSdk 31.
-* Updated to Android compileSdk 31. 
+* Updated to Android compileSdk 31.
 * Updated to Android Build Tools 31.0.0.
 * Updated to Ktlint version 0.43.0.
 * Updated to Ktlint Gradle Plugin 10.2.0.
 * Updated to Kotlin Serialization 1.3.0.
 * Updated to Detekt 1.19.0-RC1.
+* Updated to Dokka 1.6.0.
+* Updated to AtomicFu 0.17.0.
 * Updated to Realm Core 11.6.1, commit: 758d238f68fa1d16409ef0565f01c38242af5bf4.